requirements.txt
# Web framework & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0

# Database & Models
sqlmodel==0.0.22
aiosqlite==0.20.0

# gRPC (Core management)
grpcio==1.66.0
grpcio-tools==1.66.0
protobuf==5.27.2

# CLI & Utilities
typer==0.15.1
click==8.1.8
rich==13.9.0
qrcode==8.0.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
jinja2==3.1.4

# Dev Tools
ruff>=0.6.0
basedpyright>=1.17.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx>=0.27.0
